fix: 바뀐 로직 수정
